+++ packages/node/src/indexer/project.service.ts
@@ -9,17 +9,15 @@ import {
   BaseProjectService,
   StoreService,
   NodeConfig,
-  ApiService,
   IProjectUpgradeService,
+  ApiService,
 } from '@subql/node-core';
-import { EthereumBlockWrapper } from '@subql/types-ethereum';
 import { Sequelize } from '@subql/x-sequelize';
 import {
   EthereumProjectDs,
   SubqueryProject,
 } from '../configure/SubqueryProject';
-import { EthereumApi } from '../ethereum';
-import SafeEthProvider from '../ethereum/safe-api';
+import { EthereumApiService } from '../ethereum';
 import { DsProcessorService } from './ds-processor.service';
 import { DynamicDsService } from './dynamic-ds.service';
 import { UnfinalizedBlocksService } from './unfinalizedBlocks.service';
@@ -29,14 +27,14 @@ const { version: packageVersion } = require('../../package.json');
 
 @Injectable()
 export class ProjectService extends BaseProjectService<
-  ApiService<EthereumApi, SafeEthProvider, EthereumBlockWrapper[]>,
+  EthereumApiService,
   EthereumProjectDs
 > {
   protected packageVersion = packageVersion;
 
   constructor(
     dsProcessorService: DsProcessorService,
-    apiService: ApiService,
+    @Inject(ApiService) apiService: EthereumApiService,
     @Inject(isMainThread ? PoiService : 'Null') poiService: PoiService,
     @Inject(isMainThread ? Sequelize : 'Null') sequelize: Sequelize,
     @Inject('ISubqueryProject') project: SubqueryProject,
@@ -71,6 +69,6 @@ export class ProjectService extends BaseProjectService<
 
   protected onProjectChange(project: SubqueryProject): void | Promise<void> {
     // TODO update this when implementing skipBlock feature for Eth
-    // this.apiService.updateBlockFetching();
+    this.apiService.updateBlockFetching();
   }
 }