+++ packages/node/src/indexer/types.ts
@@ -1,17 +1,8 @@
 // Copyright 2020-2023 SubQuery Pte Ltd authors & contributors
 // SPDX-License-Identifier: GPL-3.0
 
-import { Entity } from '@subql/types-core';
-
-export enum OperationType {
-  Set = 'Set',
-  Remove = 'Remove',
-}
-
-export type OperationEntity = {
-  operation: OperationType;
-  entityType: string;
-  data: Entity | string;
-};
+import { EthereumBlock, LightEthereumBlock } from '@subql/types-ethereum';
 
 export type BestBlocks = Record<number, string>;
+
+export type BlockContent = EthereumBlock | LightEthereumBlock;